chore: move `style` to `markdown`
