+++ src/markdown/mod.rs
@@ -2,3 +2,4 @@ pub(crate) mod elements;
 pub(crate) mod html;
 pub(crate) mod parse;
 pub(crate) mod text;
+pub(crate) mod text_style;
